# Conflicts:
#	backend-rest-api/src/main/java/com/nitsoft/ecommerce/service/categories/CategoriesServiceImpl.java

+++ backend-rest-api/src/main/java/com/nitsoft/ecommerce/database/model/Company.java
@@ -39,7 +39,7 @@ public class Company implements Serializable {
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Basic(optional = false)
     @Column(name = "company_id")
-    private Integer companyId;
+    private long companyId;
     
     @Basic(optional = false)
     @Column(name = "name")
